refactor(github): split BatchRepoForm into reusable components
- AnalysisOptions: Extract analysis options into a separate component
- FormActions: Create a dedicated component for form actions
- ProcessingProgress: Build a standalone processing progress component
- RepoStatusItem: Design a reusable component for individual repo status
- RepoStatusList: Implement a component for listing repo statuses
